refactor(workspace): migrate skill-domain code from dcc-mcp-utils to dcc-mcp-skills
Closes #498. Moves SkillFeedback / SkillVersionManifest / SkillVersionEntry, all skill-search-path env helpers, skill-scope dirs, copy_skill_to_*_dir helpers and the related Python bindings into dedicated dcc-mcp-skills modules: constants.rs / paths.rs / feedback.rs / versioning.rs.

dcc-mcp-utils now only owns the platform-dir helpers (get_platform_dir / get_config_dir / get_data_dir / get_log_dir / get_tools_dir) and the FilesystemError type, plus a small ActionResult-related constants set.

Public Python API surface is unchanged (stubgen produces zero diff); 8391 pytest pass.
