+++ crates/dcc-mcp-server/Cargo.toml
@@ -18,7 +18,6 @@ dcc-mcp-actions.workspace = true
 dcc-mcp-skills.workspace = true
 dcc-mcp-protocols.workspace = true
 dcc-mcp-logging.workspace = true
-dcc-mcp-utils.workspace = true
 dcc-mcp-transport.workspace = true
 serde.workspace = true
 serde_json.workspace = true
