+++ crates/dcc-mcp-skills/src/loader/files.rs
@@ -1,9 +1,9 @@
 use std::path::Path;
 
-use dcc_mcp_models::SkillMetadata;
-use dcc_mcp_utils::constants::{
+use crate::constants::{
     DEPENDS_FILE, SKILL_METADATA_DIR, SKILL_SCRIPTS_DIR, is_supported_extension,
 };
+use dcc_mcp_models::SkillMetadata;
 use dcc_mcp_utils::filesystem::path_to_string;
 
 /// Enumerate files in a directory matching a filter predicate on the file extension.
