+++ crates/dcc-mcp-skills/src/watcher_filter.rs
@@ -43,7 +43,7 @@ pub(crate) fn is_skill_related(path: &Path) -> bool {
 
     // Script files (check extension against supported list)
     if let Some(ext) = path.extension().and_then(|e| e.to_str()) {
-        if dcc_mcp_utils::constants::is_supported_extension(ext) {
+        if crate::constants::is_supported_extension(ext) {
             return true;
         }
     }
