+++ src/lib.rs
@@ -222,20 +222,20 @@ fn register_utils(m: &Bound<'_, PyModule>) -> PyResult<()> {
         dcc_mcp_utils::filesystem::py_get_data_dir,
         dcc_mcp_utils::filesystem::py_get_log_dir,
         dcc_mcp_utils::filesystem::py_get_tools_dir,
-        dcc_mcp_utils::filesystem::py_get_skills_dir,
-        dcc_mcp_utils::filesystem::py_get_skill_paths_from_env,
-        dcc_mcp_utils::filesystem::py_get_app_skill_paths_from_env,
-        dcc_mcp_utils::filesystem::py_get_user_skill_paths_from_env,
-        dcc_mcp_utils::filesystem::py_get_team_skill_paths_from_env,
-        dcc_mcp_utils::filesystem::py_get_app_user_skill_paths_from_env,
-        dcc_mcp_utils::filesystem::py_get_app_team_skill_paths_from_env,
-        dcc_mcp_utils::filesystem::py_get_user_skills_dir,
-        dcc_mcp_utils::filesystem::py_get_team_skills_dir,
-        dcc_mcp_utils::filesystem::py_copy_skill_to_user_dir,
-        dcc_mcp_utils::filesystem::py_copy_skill_to_team_dir,
-        dcc_mcp_utils::filesystem::py_record_skill_feedback,
-        dcc_mcp_utils::filesystem::py_get_skill_feedback,
-        dcc_mcp_utils::filesystem::py_get_skill_version_manifest,
+        dcc_mcp_skills::paths::py_get_skills_dir,
+        dcc_mcp_skills::paths::py_get_skill_paths_from_env,
+        dcc_mcp_skills::paths::py_get_app_skill_paths_from_env,
+        dcc_mcp_skills::paths::py_get_user_skill_paths_from_env,
+        dcc_mcp_skills::paths::py_get_team_skill_paths_from_env,
+        dcc_mcp_skills::paths::py_get_app_user_skill_paths_from_env,
+        dcc_mcp_skills::paths::py_get_app_team_skill_paths_from_env,
+        dcc_mcp_skills::paths::py_get_user_skills_dir,
+        dcc_mcp_skills::paths::py_get_team_skills_dir,
+        dcc_mcp_skills::paths::py_copy_skill_to_user_dir,
+        dcc_mcp_skills::paths::py_copy_skill_to_team_dir,
+        dcc_mcp_skills::feedback::py_record_skill_feedback,
+        dcc_mcp_skills::feedback::py_get_skill_feedback,
+        dcc_mcp_skills::versioning::py_get_skill_version_manifest,
         dcc_mcp_pybridge::type_wrappers::py_unwrap_value,
         dcc_mcp_pybridge::type_wrappers::py_unwrap_parameters,
         dcc_mcp_pybridge::type_wrappers::py_wrap_value,
@@ -254,9 +254,9 @@ fn register_utils(m: &Bound<'_, PyModule>) -> PyResult<()> {
         dcc_mcp_pybridge::type_wrappers::IntWrapper,
         dcc_mcp_pybridge::type_wrappers::FloatWrapper,
         dcc_mcp_pybridge::type_wrappers::StringWrapper,
-        dcc_mcp_utils::filesystem::SkillFeedback,
-        dcc_mcp_utils::filesystem::SkillVersionManifest,
-        dcc_mcp_utils::filesystem::SkillVersionEntry,
+        dcc_mcp_skills::feedback::SkillFeedback,
+        dcc_mcp_skills::versioning::SkillVersionManifest,
+        dcc_mcp_skills::versioning::SkillVersionEntry,
         dcc_mcp_logging::python::PyFileLoggingConfig,
     );
     Ok(())
@@ -320,6 +320,7 @@ fn register_scheduler(m: &Bound<'_, PyModule>) -> PyResult<()> {
 #[cfg(feature = "python-bindings")]
 fn register_constants(m: &Bound<'_, PyModule>) -> PyResult<()> {
     use dcc_mcp_logging::constants as log_constants;
+    use dcc_mcp_skills::constants as skill_constants;
     use dcc_mcp_utils::constants;
     add_constants!(
         m,
@@ -328,13 +329,13 @@ fn register_constants(m: &Bound<'_, PyModule>) -> PyResult<()> {
         "DEFAULT_DCC"        => constants::DEFAULT_DCC,
         "DEFAULT_VERSION"    => constants::DEFAULT_VERSION,
         "DEFAULT_MIME_TYPE"  => constants::DEFAULT_MIME_TYPE,
-        "SKILL_METADATA_FILE"=> constants::SKILL_METADATA_FILE,
-        "SKILL_SCRIPTS_DIR"  => constants::SKILL_SCRIPTS_DIR,
-        "SKILL_METADATA_DIR" => constants::SKILL_METADATA_DIR,
-        "ENV_SKILL_PATHS"    => constants::ENV_SKILL_PATHS,
-        "ENV_USER_SKILL_PATHS" => constants::ENV_USER_SKILL_PATHS,
-        "ENV_TEAM_SKILL_PATHS" => constants::ENV_TEAM_SKILL_PATHS,
-        "ENV_DISABLE_ACCUMULATED_SKILLS" => constants::ENV_DISABLE_ACCUMULATED_SKILLS,
+        "SKILL_METADATA_FILE"=> skill_constants::SKILL_METADATA_FILE,
+        "SKILL_SCRIPTS_DIR"  => skill_constants::SKILL_SCRIPTS_DIR,
+        "SKILL_METADATA_DIR" => skill_constants::SKILL_METADATA_DIR,
+        "ENV_SKILL_PATHS"    => skill_constants::ENV_SKILL_PATHS,
+        "ENV_USER_SKILL_PATHS" => skill_constants::ENV_USER_SKILL_PATHS,
+        "ENV_TEAM_SKILL_PATHS" => skill_constants::ENV_TEAM_SKILL_PATHS,
+        "ENV_DISABLE_ACCUMULATED_SKILLS" => skill_constants::ENV_DISABLE_ACCUMULATED_SKILLS,
         "ENV_LOG_LEVEL"      => log_constants::ENV_LOG_LEVEL,
         "DEFAULT_LOG_LEVEL"  => log_constants::DEFAULT_LOG_LEVEL,
         "ENV_LOG_FILE"       => log_constants::ENV_LOG_FILE,
